* master:
  (maint) Remove usage of obsolete methods
  (#1000) Ensure sensitive arguments are detected
  (maint) Update to latest Chocolatey NuGet package
  (maint) Update to latest Chocolatey package
  (build) Re-enable release notes export
  (maint) Added copyright to nuspec file
  (build) Updated GRM config file
  (maint) Update to released beta packages
  (maint) Switch from int to long
  (maint) Update to latest beta Chocolatey packages
  (maint) Remove @everyone from Discord notification
  (maint) Bump to beta package in nuspec
  (maint) Update to latest beta Chocolatey packages
  (build) Update to latest Chocolatey.Cake.Recipe
  (build) Remove unnecessary configuration
  (build) Add notification templates
  (maint) Prevent download of relase notes
  (#974) Update to released beta packages

+++ Source/ChocolateyGui.Common/ViewModels/Items/IPackageViewModel.cs
@@ -23,7 +23,7 @@ namespace ChocolateyGui.Common.ViewModels.Items
 
         string Description { get; set; }
 
-        int DownloadCount { get; set; }
+        long DownloadCount { get; set; }
 
         string GalleryDetailsUrl { get; set; }
 
@@ -73,7 +73,7 @@ namespace ChocolateyGui.Common.ViewModels.Items
 
         NuGetVersion Version { get; set; }
 
-        int VersionDownloadCount { get; set; }
+        long VersionDownloadCount { get; set; }
 
         Task Install();
